pyproject.toml
[project]
name = "8-chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=1.2.15",
    "langchain-community>=0.4.1",
    "langchain-ollama>=1.1.0",
    "langsmith>=0.7.34",
    "python-dotenv>=1.2.2",
    "streamlit>=1.56.0",
]
